Relative log abundance (RLA) plot

The RLA function is a wrapper function to help easily visualise normalisation results on a data set. It is equivalent to the reltaive log expression (RLE) plot under the gene expression contexts.

RLA(
  data,
  smpID = NULL,
  repCol = NULL,
  repLabel = NULL,
  smpName = NULL,
  ylim = NULL,
  guides = NULL,
  title = NULL
)

Arguments

data
The m-by-n data matrix with samples on the rows and measurements ( proteins etc.) on the columns. Missing data (NAs) are allowed.

smpID
Sample IDs of the data. Default is NULL. This argument is useful when one wishes to display the samples in an order different from the sample orders of input data.

repCol
The replication color codes. Useful when one wishes to colour the RLA boxes to indicate replications. Default is NULL.

repLabel
Color legend names.

smpName
Sample names. Sample names to display on the axis. Default is NULL.

ylim
The y-axis labels. A numeric vector of length 2.

guides
Numeric. Horizontal dashed lines for reference.

title
Character. Title of the output plot.

Value

The RLA plot.

References

Gandolfo, L. C., & Speed, T. P. (2018). RLE plots: Visualizing unwanted variation in high dimensional data. PloS one, 13(2), e0191629.